fix: make itemCount 64bit (#520)
* fix: make itemCount 64bit

* test: fix

* fix: lint

+++ internal/common/analyze.go
@@ -11,7 +11,7 @@ import (
 // CurrentProgress struct
 type CurrentProgress struct {
 	CurrentItemName string
-	ItemCount       int
+	ItemCount       int64
 	TotalSize       int64
 }
